Move onboarding form to the website (#2017)
- [x] Make repo-ingestion fault tolerant (deploy on both wato-aks1 and
wato-kube)
- [x] Set up utility functions from the old form like hashing passwords.
- [x] Render markdown in field descriptions.
- [x] Add "Advanced" dropdown: download config (show error in dialog,
have a button to download anyway), save for later (useful for admins
creating templates). Can use the [Dropdown
Menu](https://ui.shadcn.com/docs/components/dropdown-menu) component
- [x] Prompt when trying to leave the page with modifications
- [x] CI-based checklist
https://github.com/WATonomous/infra-config/pull/2553
- [x] Resolve the 500 error when no file is changed:
https://github.com/WATonomous/infra-config/pull/2017#issuecomment-2028562464
(in a separate PR) https://github.com/WATonomous/infra-config/pull/2552
- [x] Form instructions
- [x] Add deprecation notice to the old form. Link the new form, also
transfer query params.
https://github.com/WATonomous/infra-config/pull/2017/commits/3b291ae374a1ef0c9bc22860e7d819b232d5e95b
- [x] Update website URLs to point to the new page.

---------

+++ scripts/generate-mdx-strings.py
@@ -1,7 +1,8 @@
 import json
-import typer
-import hashlib
+from itertools import chain
 from pathlib import Path
+
+import typer
 
 app = typer.Typer()
 
@@ -51,21 +52,41 @@ def get_all_strings(json_file_path):
     return strings
 
 @app.command()
-def dump_mdx(strings: list[str], output_dir: str):
+def dump_mdx(strings: list[str], output_dir: str, overwrite: bool = False):
     output_dir = Path(output_dir)
-    Path.mkdir(output_dir, exist_ok=True)
+    if output_dir.exists() and not overwrite:
+        raise Exception(f"Output directory '{output_dir}' already exists. Use --overwrite to overwrite it.")
+    Path.mkdir(output_dir, parents=True)
 
+    hash_to_string = {}
     for s in strings:
-        basename = f"{hash_code(s)}.mdx"
+        h = hash_code(s)
+        if h in hash_to_string and hash_to_string[h] != s:
+            raise Exception(f"ERROR: Hash collision: '{s}' and '{hash_to_string[h]}' have the same hash code {h}")
+        hash_to_string[h] = s
+
+    for h, s in hash_to_string.items():
+        basename = f"{h}.mdx"
 
         with open(output_dir / basename, "w") as file:
             file.write(s)
 
+    strings_file = output_dir / "strings.ts"
+    with open(strings_file, "w") as file:
+        for h in hash_to_string.keys():
+            file.write(f"import String{h} from './{h}.mdx'\n")
+        file.write("\n")
+
+        file.write("export default {\n")
+        for h in hash_to_string.keys():
+            file.write(f"  '{h}': String{h},\n")
+        file.write("}\n")
+
     print(f"Dumped {len(strings)} strings to {output_dir}")
 
 @app.command()
-def json_to_mdx(json_file_path: str, output_dir: str):
-    strings = get_all_strings(json_file_path)
+def json_to_mdx(json_file_paths: list[str], output_dir: str):
+    strings = list(chain.from_iterable(get_all_strings(p) for p in json_file_paths))
     dump_mdx(strings, output_dir)
 
 if __name__ == '__main__':
